Add system test for different exercise statuses
