+++ app/helpers/react_components/student/open_editor_button.rb
@@ -28,7 +28,8 @@ module ReactComponents
       memoize
       def links
         {
-          start: Exercism::Routes.start_track_exercise_path(exercise.track, exercise)
+          start: Exercism::Routes.start_track_exercise_path(exercise.track, exercise),
+          exercise: Exercism::Routes.edit_track_exercise_path(exercise.track, exercise)
         }
       end
